pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "notion2hugo"
version = "0.1.0"
description = "Export Notion database to Hugo markdown"
authors = [{ name = "Hendra Manudinata", email = "manoedinata@gmail.com" }]
requires-python = ">=3.10"
dependencies = [
    "notion-client",
    "python-dotenv",
    "pyyaml",
]

[project.scripts]
notion2hugo = "notion2hugo:main"
